Refactor and add to `maputils.py`
Add the `create_geojson` function to maputils. This takes in the
dictionary generated by `site_dict` and outputs a dictionary formatted
as geojson.

Change `write_geojson` to use `create_geojson` and the builtin json
library to write to a file

Add `projection_info` function to retrieve proj4 the projection
definition string on a projection given its epsg code. The api endpoint
to be used with this is in `config.py`