Add basic sites map on url `/map/`
Add template `sitemap,html`, which uses d3.js and proj4js to render
dots representing the specified sites.

In `views.py` add `sitemap`, which calls to `maputils` and renders the
sitemap template.

In `config.py` add entries for the spatialreference.org template url and
the epsg code of the map projection to be used. Both entries are used in
calls to `maputils`.