Add scale as a factor and check for valid EPSG code
In `views.py` add a `scale` variable to be passed to d3. This allows
indirect adjustment of the bounding box. Change `gage_map.js` to use
this variable.

In `map_utils.py` in `projection_info` add an assertion that the request
returns a 200 status code, i.e., that the EPSG code given is valid.