Switch to load rivers from file
Add `rivers.json` to the `static/geojson` directory.
In views load this file and package it with the other map data.
In map utils, remove `fetch_rivers` in favor of loading from file.

In `gage_map.js` add logic to display rivers on the map.
Rearrange colors in `mapstyle.css` to allow rivers to be blue.

Update `config.py` to include current map settings. (set to None or the
empty string)