Link map click to add/remove hydrograph lines
In `hydrograph.js`:

Add the `change_lines` function, which takes a list of site ids to
display, and returns a subset of the original data containing only
data from those sites.

Make `display_ids` private, settable only through calls
to `change_lines`, and readable through a getter.
Use `display_ids` to keep track of which lines are being displayed both
for actions within the hydrograph and interactions with the map.

Alter the `click` function to take only the datapoint as a parameter,
and then call `change_lines` with an updated version of `display_ids`.

In `map.js`:

Add new function `addaccent` to highlight a gage on the map based on
the given id.

Add `click` function which takes a site id and toggles display of that
site's data on the hydrograph.