Change `add_circles` to generalize adding ID
Add parameter `property_for_id` to be used to specify the field in the
`properties` of each feature that should be used to give each circle its
id. If the parameter is null or not provided at all,
or if the feature has no such property, no id will be given.