Switch to d3 drag interaction
Switch from using event handlers on mousedown, mousemove, and mouseup
attached to the map svg to using the d3 drag behavior, with the events
'start', 'drag', and 'end'. Set both the background and the rivers
groups to call the drag behavior.

Add logic to distinguish drags from errant clicks, and ignore the
latter.