Add doubleclick interaction to reset hydrograph
Alter the click listener on the hydrograph to use a custom doubleclick
implementation, which I don't think is any good.

Add `default_display_ids` to store the original set of displayed ids.
Set this to a deep copy of `FV.hydrograph_display_ids` in `init`.
Alter `change_lines` to deep copy to avoid overwriting defaults.