Bind dblclick to svg, not voronoi
Bind the dblclick action to the svg, instead of the voronoi.
This fires two `click` actions before each `dblclick`, but I haven't
found a way to avoid this that isn't buggy.

Add a rectangle to the background of the hydrograph so that a dblclick
on a blank graph restores the default lines. Before, only clicking on
the axes would trigger an event.