Begin using state to track more complex gage style
Add gage styles `bland` and `accent-bland`, which are less opaque than
the normal styles. When one gage is moused over, all other gages will be
given the bland version of their style.

To track these changes, I am making significant changes to `state`.
`state.gages` is now an object, not an array. Each key in `state.gages`
is a gage id, and its value is an object containing information about
that gage.

I'm in the middle of making these changes.